Feat(.welcome): header/footer 컴포넌트화
