+++ assets/css/include.css
@@ -31,8 +31,6 @@
   margin: 0 20px 10px;
   padding: 0 10px;
   /* border: 2px solid pink; */
-  text-decoration: none;
-  /* Remove default underline */
 }
 
 .main-header .main-nav .main-nav__items .main-nav__item a::before {
