+++ components/footer/FooterComponent.vue
@@ -1,0 +1,30 @@
+<template>
+  <div class="wrap">
+    <footer>
+      <ul>
+        <li>COPi</li>
+        <li><a href="/">COP<span>i</span>소개</a></li>
+        <li><a href="/">이용약관</a></li>
+        <li><a href="/">개인정보처리방침</a></li>
+        <li><a href="/sum">광고상품소개</a></li>
+      </ul>
+      <div class="copyright">©COPI. ALL RIGHTS RESERVED</div>
+      <a id="goTopBtn" href="#top" class="go-top"><i class="fa-regular fa-circle-up"></i></a>
+    </footer>
+  </div>
+</template>
+
+<script setup lang="ts"></script>
+
+<style scoped>
+/* 전체 배경 */
+/* .wrap {
+  background: #1b1b1b;
+} */
+
+header,
+footer {
+  position: relative;
+  z-index: 1;
+}
+</style>
